Add Ansible lint error [106] to skip list
ansible-lint generates the folowing error:

[106] Role name coglinev3.ansible_python does not match
``^[a-z][a-z0-9_]+$`` pattern

ansible-lint does not allow a dot between the author name and the
role name. However, the naming convention for Ansible Galaxy is
author_name.role_name. I think this is an ansible-lint bug. So I added
the error message [106] to the list of ignored errors in the
.ansible-lint file.